+++ cli/tests/test_show_command.rs
@@ -270,6 +270,9 @@ fn test_show_with_no_template() {
 
     For more information, try '--help'.
     Hint: The following template aliases are defined:
+    - builtin_config_item
+    - builtin_config_list
+    - builtin_config_list_detailed
     - builtin_log_comfortable
     - builtin_log_compact
     - builtin_log_compact_full_description